Make Clippy happy
